# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Proyectos de Pavimento",
        "description": "Documentación de la API para la gestión de proyectos, abscisas, losas y patologías. Creada para la aplicación móvil y web.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projects/create": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Crear un nuevo proyecto",
                "description": "create a new project",
                "operationId": "e095ad0090afd79434a043704628ab0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proyecto Vía Central"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Rehabilitación de vía principal"
                                    },
                                    "image": {
                                        "description": "Imagen del proyecto",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "start_lat": {
                                        "type": "string",
                                        "example": "4.60971"
                                    },
                                    "start_lng": {
                                        "type": "string",
                                        "example": "-74.08175"
                                    },
                                    "end_lat": {
                                        "type": "string",
                                        "example": "4.70000"
                                    },
                                    "end_lng": {
                                        "type": "string",
                                        "example": "-74.10000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto creado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Obtener todos los proyectos del usuario autenticado",
                "description": "get all projects",
                "operationId": "5995c33f440f15317c83bd7ddc72f90d",
                "responses": {
                    "200": {
                        "description": "Lista de proyectos paginada"
                    },
                    "404": {
                        "description": "No se encontraron proyectos"
                    }
                }
            }
        },
        "/api/projects/details": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Obtener detalles de un proyecto por ID",
                "description": "get details project by id",
                "operationId": "a84e2e09107e09ddec26b5b56d4e9769",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del proyecto"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Obtener detalles de un proyecto por parámetro de ruta",
                "description": "get details project by id from route parameter",
                "operationId": "2da628cf1021d0feb72651c8b0d05833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del proyecto"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/projects/collaborator/register": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Registrar un colaborador en un proyecto",
                "description": "register a new collaborator into a project",
                "operationId": "556429227a1d29305dea2f2d7a8a2efc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Colaborador registrado con éxito"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/projects/all": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Obtener todos los proyectos (limitado)",
                "description": "get all project",
                "operationId": "ba3d5c19a6cccacaa7d5b6d290ad5f1b",
                "responses": {
                    "200": {
                        "description": "Lista de proyectos"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/model/ia": {
            "post": {
                "tags": [
                    "Model"
                ],
                "summary": "Consultar el modelo de IA con una imagen",
                "description": "Consulta el modelo de IA",
                "operationId": "63914d2694fee6d8643d039064c0cd68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Imagen para procesar en el modelo IA",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta del modelo IA"
                    },
                    "400": {
                        "description": "Error de validación o imagen no válida"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por Bearer Token (Laravel Sanctum). Incluir el token en el header 'Authorization'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}